[probes.external] Refactor the external probe code
